tests: have one reactive test
