+++ __tests__/reactive.test.ts
@@ -196,7 +196,7 @@ describe('bindVertex reactive wrapper', () => {
 
     // Transient edits first
     const when = new Date('2025-01-03T00:00:00.000Z');
-    person.useTransient(p => {
+    person.$useTransient(p => {
       p.name = 'Draft' as any;
       p.age = 25 as any;
       p.createdAt = when as any;
@@ -214,7 +214,7 @@ describe('bindVertex reactive wrapper', () => {
     expect(tree.getVertexProperty(v.id, 'age', false)).toBeUndefined();
 
     // Promote transients -> persist them
-    person.commitTransients();
+    person.$commitTransients();
     expect(tree.getVertexProperty(v.id, '_n', false)).toBe('Draft');
     expect(tree.getVertexProperty(v.id, 'age', false)).toBe(25);
     // createdAt persisted as ISO
@@ -395,7 +395,7 @@ describe('bindVertex reactive wrapper', () => {
     // Add children
     const child1 = tree.newVertex(parent.id);
     const child2 = tree.newVertex(parent.id);
-    
+
     // Wait for batched events to process (~33ms)
     await new Promise(resolve => setTimeout(resolve, 50));
 
@@ -408,4 +408,33 @@ describe('bindVertex reactive wrapper', () => {
     // Cleanup
     unobserve();
   });
+
+  test('sync between Vertex and bound proxy', () => {
+    const tree = new RepTree('peer1');
+    const root = tree.createRoot();
+
+    // Create a child vertex normally (non-bound)
+    const v = root.newChild();
+
+    // Create two bound proxies pointing at the same vertex
+    const refA = bindVertex<{ score: number }>(tree, v.id);
+    const refB = bindVertex<{ score: number }>(tree, v.id);
+
+    // 1) Write via Vertex API
+    v.setProperty('score', 10);
+
+    // Reads via bound proxies should reflect the value
+    expect(refA.score).toBe(10);
+    expect(refB.score).toBe(10);
+
+    // 2) Write via a bound proxy
+    refA.score = 42;
+
+    // Vertex.getProperty should reflect the new value
+    expect(v.getProperty('score')).toBe(42);
+
+    // Other bound proxies also see the change
+    expect(refB.score).toBe(42);
+  });
+
 });
